requirements/pypi.txt
# Framework
django

# Command
django-click

# REST
djangorestframework
requests

# Other
django-filter
